requirements.txt
altgraph==0.17.4
certifi==2025.1.31
charset-normalizer==3.4.1
getmac==0.9.5
idna==3.10
packaging==24.2
pefile==2023.2.7
psutil==7.0.0
py-cpuinfo==9.0.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==78.1.0
urllib3==2.3.0